# fix: faster parse options