--- a/test/internal/parse-options.js
+++ b/test/internal/parse-options.js
@@ -18,29 +18,26 @@ t.test('truthy non-objects always loose mode, for backwards comp', t => {
   t.end()
 })
 
-t.test('objects only include truthy flags we know about, set to true', t => {
-  t.strictSame(parseOptions(/asdf/), {})
-  t.strictSame(parseOptions(new Error('hello')), {})
-  t.strictSame(parseOptions({ loose: true, a: 1, rtl: false }), { loose: true })
+t.test('any object passed is returned', t => {
+  t.strictSame(parseOptions(/asdf/), /asdf/)
+  t.strictSame(parseOptions(new Error('hello')), new Error('hello'))
+  t.strictSame(parseOptions({ loose: true, a: 1, rtl: false }), { loose: true, a: 1, rtl: false })
   t.strictSame(parseOptions({ loose: 1, rtl: 2, includePrerelease: 10 }), {
-    loose: true,
-    rtl: true,
-    includePrerelease: true,
+    loose: 1,
+    rtl: 2,
+    includePrerelease: 10,
   })
   t.strictSame(parseOptions({ loose: true }), { loose: true })
   t.strictSame(parseOptions({ rtl: true }), { rtl: true })
   t.strictSame(parseOptions({ includePrerelease: true }), { includePrerelease: true })
   t.strictSame(parseOptions({ loose: true, rtl: true }), { loose: true, rtl: true })
-  t.strictSame(parseOptions({ loose: true, includePrerelease: true }), { loose: true, includePrerelease: true })
-  t.strictSame(parseOptions({ rtl: true, includePrerelease: true }), { rtl: true, includePrerelease: true })
-  t.end()
-})
-
-t.test('should skip validation when options is already parsed', t => {
-  const options = { loose: true, rtl: true }
-  const parsedOptions = parseOptions(options)
-
-  t.equal(parseOptions(parsedOptions) === parsedOptions, true)
-  t.not(parseOptions(options) === parsedOptions, false)
+  t.strictSame(parseOptions({ loose: true, includePrerelease: true }), {
+    loose: true,
+    includePrerelease: true,
+  })
+  t.strictSame(parseOptions({ rtl: true, includePrerelease: true }), {
+    rtl: true,
+    includePrerelease: true,
+  })
   t.end()
 })