@H4ad 2023-04-06
Just to bring more attention, @ljharb points out a very interesting thing that could affect this PR at https://github.com/npm/node-semver/pull/536#pullrequestreview-1374076868.

 If the assumption is true, we can clean all the `varN` variables and just return the original object options.
This will only affect the tests, so we will need to fix them.