@jakebailey 2023-04-06
Here's my DT stress test using pnpm. Before:

```
Done in 2m 14.5s
total time:  134.82s
user time:   162.81s
system time: 31.61s
CPU percent: 144%
max memory:  1752 MB
```

This PR (`"semver": "github:h4ad-forks/node-semver#fix/faster-parse-options"`):

```
Done in 1m 42.2s
total time:  102.43s
user time:   115.96s
system time: 30.80s
CPU percent: 143%
max memory:  1774 MB
```

Which according to https://how-much-faster.glitch.me/ is:

- 31.62% faster
- 1.32x faster
- 24.02% less time