@wraithgar 2023-04-06
Stepping back and broadly asking the question "what is `parseOptions` for"? I think we have the following:

1. standardize object for memoization
   - not needed thanks to #536)
2. coercion of `true|false` into `{ loose: true:false }`
   - required, semver major change to remove this
3. coercion of empty options into `{}`
   - not absolutely required, but definitely a best practice. Relying on the fact that `true` doesn't have a `loose` attribute is quite the antipattern
4. filtering out of unknown options
   - we already don't do anything if extra options are passed.
   - potentially a memory savings though?  Is the tradeoff worth it for the extra cpu time?

If all we did were `2` and `3` I think we would be ok?