@jakebailey 2023-04-06
Personally, if I were to rewrite this library from scratch (something I'm trying to do to explore perf improvements, though I don't know if I will actually publish it), I wouldn't be storing these options in `SemVer` at all.

Looking at `SemVer`, only two options are used, `loose` and `includePrerelease`.

`loose` means that the parser can accept things like leading zeros. But once it's parsed, the loose-ness isn't important anymore. If it's being compared with another version, that's a different version, and the loose-ness of a differnet version is not important.

`includePrerelease`, on the other hand, isn't used at all. That option really only makes sense for `Range`. And even then, I still think that it's not a _property_ of a `Range`; it's a property of `satisfies` and other helpers and how they decide to _interpret_ a `Range`.

So to me, it feels like an equivalently good change to this PR would be to just check `loose` once, store it as a prop for people to know if a `SemVer` came from a loose parse, but otherwise not try and do anything.