@wraithgar 2023-04-06
`rtl` is used in `coerce` and only there, and that `coerce` is not used by anything else in semver.