@H4ad 2023-04-06
Well, I think we agree to keep the object as-is and don't do any modification, and just validate the cases of `undefined` and `boolean`.

In this case, I will create a new commit instead of modifying the first one with the test modified and the new behavior.