@wraithgar 2023-04-06
> Also, this can be considered as breaking change?

It depends on what the changes needed are.  Just cause our tests have to change doesn't necessarily mean it's a breaking change.  If folks are going to get different results after this change by passing the exact same things as before, then it is a breaking change.