@wraithgar 2023-04-06
I think once the linting errors are fixed we are in a good place here.

@H4ad how do you feel about this PR in contrast to your other idea of using bitmasks?  Do you still think that is worth exploring?