@wraithgar 2023-04-06
Like #536 I'm giving this a 👍 and letting it sit through the weekend to let folks catch up.

This PR should definitely *not* land before #536, it is predicated off the fact that the parsed options are no longer being used to memoize a cache.